feat: add autocompleteHelpers
